bots/spike-detector/requirements.txt
python-socketio[client]==5.11.0
websocket-client==1.6.4
requests==2.31.0
python-engineio==4.9.0